# src/main/resources/openapi/itemtree-api.yaml
openapi: 3.0.3
info:
  title: ItemTree API
  description: REST service fronting the ITEMTREE Oracle table.
  version: 1.0.0
...
paths:
...
  /api/v1/itemtree/tree:
    get:
      tags: [tree]
      operationId: getTree
      summary: Get trimmed tree view for the resolved user (skeleton + ancestor chain + home-folder children)
      parameters:
        - $ref: '#/components/parameters/XIceUser'
        - $ref: '#/components/parameters/XImpersonatedUser'
      responses:
        '200':
          description: Flat list of ItemNode each with path populated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemNode'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/itemtree/tree/{rootId}/subtree:
    get:
      tags: [tree]
      operationId: getSubtree
      summary: Get root + immediate children of rootId as a flat list with paths
      parameters:
        - name: rootId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XIceUser'
        - $ref: '#/components/parameters/XImpersonatedUser'
      responses:
        '200':
          description: Flat list of ItemNode (root first, then immediate children) each with path populated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemNode'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/itemtree/tree/{rootId}/subtree-full:
    get:
      tags: [tree]
      operationId: getSubtreeFull
      summary: Get the full recursive subtree rooted at rootId as a flat list with paths
      parameters:
        - name: rootId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XIceUser'
        - $ref: '#/components/parameters/XImpersonatedUser'
      responses:
        '200':
          description: Flat list of ItemNode (root + all descendants in BFS order) each with path populated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemNode'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  parameters:
    XIceUser:
      name: X-Ice-User
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 20
    XImpersonatedUser:
      name: X-Impersonated-User
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 20
  schemas:
    ItemNode:
      type: object
      required: [itemTreeId, parentId, name, type, lastUpdate, lastUpdateUser]
      properties:
        itemTreeId:
          type: integer
          format: int64
        parentId:
          type: integer
          format: int64
          description: "Always populated; 0 for root's parent"
        name:
          type: string
        type:
          type: string
        path:
          type: string
          nullable: true
          description: "Root-anchored slash-separated path, e.g. root/Folder1/IceReport; populated only on /tree and /subtree responses"
        lastUpdate:
          type: string
          format: date-time
        lastUpdateUser:
          type: string
...
    Problem:
      type: object
      description: RFC 7807 problem detail
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        errorCode:
          type: string
          description: "Machine-readable code e.g. PARENT_NOT_FOUND, MOVE_INTO_DESCENDANT"
        traceId:
          type: string
  responses:
...
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ServiceUnavailable:
      description: Service unavailable — cache not ready
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'